+++ rules/astro.js
@@ -1,7 +1,0 @@
-module.exports = {
-  customSyntax: "postcss-html",
-  rules: {
-    // Allow javascript custom properties (camelCase)
-    "custom-property-pattern": null,
-  },
-};